requirements.txt
Flask==2.3.3
anthropic==0.23.1
python-dotenv==1.0.0
gunicorn==21.2.0
google-cloud-storage>=2.10.0
geopandas>=0.12.0
shapely>=2.0.0
pandas==2.0.3
requests==2.31.0
google-auth==2.23.0
fiona==1.9.4
pyproj==3.6.0
uuid==1.30
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
google-cloud-bigquery>=3.11.0
numpy>=1.24.0
google-cloud-storage>=2.10.0
google-auth>=2.22.0


# Additional utilities (if needed)
# Uncomment these if your analysis modules use them
# matplotlib>=3.7.0
# folium>=0.14.0
# contextily>=1.3.0